音領域のクラス更新
