+++ pro_builder_test/Assets/ProgrammerSample/KazuoFile/Prefab/SoundAreaSpawner.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3447c55d0811648f89060e4730a85357
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
