+++ pro_builder_test/Assets/ProgrammerSample/KazuoFile/Script/Player/SoundAreaSpawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ce091a8708ec447249f8b87f44d9b8f4
+guid: 765db4dbc20034c73b64e234a0994dfa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
